+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Many changes to .gitconfig
+* Bootstrap preserves symlinks for etc and .vim
+* bootstrap excludes CHANGELOG
 
 ### Fixed